pyproject.toml
[project]
name = "nebulous-sh"
version = "0.1.131"
description = "A globally distributed container runtime"
readme = "README.md"
requires-python = ">=3.10.14"
dependencies = [
    "boto3>=1.37.30",
    "dill>=0.3.8",
    "loguru>=0.7.3",
    "openai>=1.68.2",
    "pillow>=10.4.0",
    "pydantic>=2.10.6",
    "pysocks>=1.7.1",
    "pyyaml>=6.0.2",
    "rclone-python>=0.1.21",
    "redis>=5.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "chatmux>=0.1.4",
    "datamodel-code-generator>=0.28.5",
    "ipykernel>=6.29.5",
    "nbformat>=5.10.4",
    "notebook>=7.3.3",
    "peft>=0.15.1",
    "pytest>=8.3.5",
    "qwen-vl-utils>=0.0.10",
    "torchvision>=0.21.0",
    "trl>=0.16.1",
    "wandb>=0.19.9",
]

[tool.pyright]
reportUnknownParameterType = false
reportMissingTypeArgument = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
